requirements.txt
fastapi
uvicorn
python-multipart
aiohttp
openai
python-dotenv
pydantic